Predict Method for a 'glm_betaselect' Object

Compute the predicted values in a model fitted by glm_betaselect().

Usage

# S3 method for glm_betaselect
predict(
  object,
  model_type = c("beta", "standardized", "raw", "unstandardized"),
  newdata,
  ...
)

Arguments

object

A glm_betaselect-class object.

model_type

The model from which the the predicted values are computed. For "beta" or "standardized", the model is the one after selected variables standardized. For "raw" or "unstandardized", the model is the one before standardization was done.

newdata

If set to a data frame, the predicted values are computed using this data frame. The data must be unstandardized. That is, the variables are of the same units as in the data frame used in glm_betaselect(). If model_type is "beta" or "standardized", it will be standardized using the setting of to_standardize when object is created in glm_betaselect().

...

Arguments to be passed to stats::predict.glm(). Please refer to the help page of stats::predict.glm().

Value

It returns the output of stats::predict.glm().

Details

It simply passes the model before or after selected variables are standardized to the predict-method of a glm object.

IMPORTANT

Some statistics, such as prediction or confidence interval, which make use of the sampling variances and covariances of coefficient estimates may not be applicable to the models with one or more variables standardized. Therefore, they should only be used for exploratory purpose.

See also

Author

Shu Fai Cheung https://orcid.org/0000-0002-9871-9448

Examples

data_test_mod_cat$p <- scale(data_test_mod_cat$dv)[, 1]
data_test_mod_cat$p <- ifelse(data_test_mod_cat$p > 0,
                              yes = 1,
                              no = 0)
logistic_beta_x <- glm_betaselect(p ~ iv*mod + cov1 + cat1,
                                  data = data_test_mod_cat,
                                  family = binomial,
                                  to_standardize = "iv")

predict(logistic_beta_x)
#>           1           2           3           4           5           6 
#> -0.14978605 -0.08134324 -0.17234895 -0.96500507 -0.41140099 -3.66500876 
#>           7           8           9          10          11          12 
#>  0.45641136  3.21533738  1.27031906  4.19968995 -1.12443124 -2.48912251 
#>          13          14          15          16          17          18 
#> -0.95030270  1.40075560 -3.87061892 -1.22196386  1.64793248  0.26351062 
#>          19          20          21          22          23          24 
#>  0.72796937 -3.10347290  2.51888487  2.95925773 -0.07194538  4.00420785 
#>          25          26          27          28          29          30 
#>  4.32166278  3.61108912  0.19253103  0.70736943  4.93483135  1.63179563 
#>          31          32          33          34          35          36 
#> -1.53121733  0.88326050 -1.56874072 -1.29975147 -2.31172710  0.80970723 
#>          37          38          39          40          41          42 
#> -2.74222060  2.76830583  0.14277124 -1.07515082  1.21997917 -3.56021526 
#>          43          44          45          46          47          48 
#> -0.37411008 -0.73927584 -0.32588437  2.92756024  3.48586073  0.79603101 
#>          49          50          51          52          53          54 
#>  0.13479416  2.15517198 -1.30365486 -0.97602940 -0.58075477 -0.10698828 
#>          55          56          57          58          59          60 
#> -2.79198254  2.97961397  0.34591136  0.64047900  2.92044253 -0.61555969 
#>          61          62          63          64          65          66 
#> -2.78373829 -3.17248201  1.04435794 -2.13417300 -3.12605466  1.46695915 
#>          67          68          69          70          71          72 
#> -0.27629999  0.06464732 -1.74345959  1.55472572  0.08787974 -3.95342379 
#>          73          74          75          76          77          78 
#> -4.48141414 -1.46129296  0.54234905  0.82867871 -0.32387428 -2.76947603 
#>          79          80          81          82          83          84 
#> -3.28820957 -1.10684008 -1.59519351 -2.15464871  1.14737887  1.12557393 
#>          85          86          87          88          89          90 
#>  2.75849747 -3.73809474 -1.33946564 -0.43821183  0.08230553 -2.66283963 
#>          91          92          93          94          95          96 
#> -2.22750598  2.04334594 -0.26442030 -0.17588281 -2.40653927  2.12545793 
#>          97          98          99         100         101         102 
#> -1.65670574  0.34246420  1.84471157 -3.41535617 -2.01583249  3.27498398 
#>         103         104         105         106         107         108 
#>  0.74220379  0.70309052 -1.71766823  0.08033343 -2.07154702 -1.00077386 
#>         109         110         111         112         113         114 
#> -2.23924613  2.67443454 -1.06628380 -0.25772980  1.66808167  1.44032300 
#>         115         116         117         118         119         120 
#> -1.19493355  0.60859601  0.14593548  1.04389437 -0.51851066  1.52205280 
#>         121         122         123         124         125         126 
#>  0.45109349  0.12090314 -3.46088450 -1.33896658  2.09200021  2.66663307 
#>         127         128         129         130         131         132 
#> -0.61582272 -2.45552247  1.05135084 -0.16159274  1.43234487  3.03601676 
#>         133         134         135         136         137         138 
#> -0.37651683  0.10070517  1.96454151  0.81307308  2.15618538  4.31247034 
#>         139         140         141         142         143         144 
#> -2.15053058 -2.52292605  2.75028684  1.73362226 -2.09302975 -2.60597462 
#>         145         146         147         148         149         150 
#>  0.72538904 -3.11278912  1.72165906 -1.64042493 -0.77554629  2.39461855 
#>         151         152         153         154         155         156 
#> -1.70863446 -0.17428506 -0.82468356  0.06074132 -1.16490507  0.29509604 
#>         157         158         159         160         161         162 
#> -0.68398793 -2.05277591  2.61431096  2.10334114 -0.52508101  3.01338466 
#>         163         164         165         166         167         168 
#>  0.11406477  2.77784767  1.54997586 -1.10641586 -0.44918478  2.91835191 
#>         169         170         171         172         173         174 
#>  1.47430119  0.99684868  0.49947626  3.12005526  0.48128938  2.77794721 
#>         175         176         177         178         179         180 
#>  0.16595507  0.16861120 -0.22376824 -1.31872945  1.85209254 -2.54602319 
#>         181         182         183         184         185         186 
#> -2.53859081 -1.88972430 -3.76687457  3.70116151 -3.21395851  2.89085930 
#>         187         188         189         190         191         192 
#> -1.40479191 -4.29621824  0.67899091  1.89972660  0.24971668 -1.79169593 
#>         193         194         195         196         197         198 
#> -2.12379832 -0.58494100 -0.27143124  0.32580796  4.84565233 -1.31177570 
#>         199         200         201         202         203         204 
#> -3.06889586  0.60330544 -0.92431558 -1.37087778  0.18234070 -0.23410712 
#>         205         206         207         208         209         210 
#> -1.72683579  2.97224220 -3.83487255 -1.59774187  3.09591749 -2.83625462 
#>         211         212         213         214         215         216 
#>  0.57501931 -1.18626605 -1.29232506  2.66235630  2.54333553  2.23068544 
#>         217         218         219         220         221         222 
#> -1.86052114  2.73540792  1.60009735  0.40208225 -2.76027561 -0.66387341 
#>         223         224         225         226         227         228 
#> -0.83102897 -1.67344023  4.13435098 -1.31790519 -0.40528978 -2.10472583 
#>         229         230         231         232         233         234 
#> -1.35923972  2.74534028  1.67655717  2.31455720  4.87817883  0.99003311 
#>         235         236         237         238         239         240 
#> -0.93981548  1.75345796 -1.75439883 -2.22564108  0.07093593  0.37916963 
#>         241         242         243         244         245         246 
#> -1.48224979  3.21970629 -1.23280102 -3.86674391  0.61502781  1.50965136 
#>         247         248         249         250         251         252 
#> -0.30348683 -2.50054556  1.37955219 -2.26703558 -2.55422215  1.72805377 
#>         253         254         255         256         257         258 
#> -0.49886308 -0.40496537  1.60535580 -2.34509376 -1.56954013 -2.76911255 
#>         259         260         261         262         263         264 
#> -1.84182565  2.11649198 -6.15055823 -0.55362932 -2.27302723  5.72724308 
#>         265         266         267         268         269         270 
#> -0.76130330 -3.99726268  0.32848350  1.01031728  0.83868488 -2.13292685 
#>         271         272         273         274         275         276 
#>  2.21734020  1.99475004 -1.56410453  1.31986166 -0.24876337  3.83256471 
#>         277         278         279         280         281         282 
#>  0.77100504  2.71582991  2.42387504  1.64199526  1.70696492 -3.21105147 
#>         283         284         285         286         287         288 
#> -2.34701803 -0.04934355  3.61706269  0.52696545  0.09540556 -4.28534988 
#>         289         290         291         292         293         294 
#> -1.82039761 -2.51962413 -0.24888060 -0.11925439 -2.03556616  2.81106731 
#>         295         296         297         298         299         300 
#> -1.50652775 -1.65268211 -0.22805644 -1.02193963 -0.06059586  1.20082000 
#>         301         302         303         304         305         306 
#>  4.01727763  0.56872269  4.78872141 -1.31901837 -0.63426880  1.89050136 
#>         307         308         309         310         311         312 
#> -1.72599666  4.27074031  0.51840777 -1.72449479 -3.26146144  1.08686025 
#>         313         314         315         316         317         318 
#>  0.81644387  1.71325127 -1.93810249 -2.65080675 -0.79585912 -3.49791403 
#>         319         320         321         322         323         324 
#>  2.40564389 -0.41507462  5.37568060 -1.45458564 -1.06606591 -0.63344740 
#>         325         326         327         328         329         330 
#>  2.88906242  0.72274356 -2.87207798 -1.11506590  0.73601529 -1.64994148 
#>         331         332         333         334         335         336 
#>  2.62882867 -2.05228099  1.62367184 -0.40381925  2.19622707 -0.41270011 
#>         337         338         339         340         341         342 
#>  1.34802157  0.74418499  1.98198510  1.90776354 -0.97432347 -0.18827959 
#>         343         344         345         346         347         348 
#> -1.23045646  1.78320427 -1.10338674 -3.36752264  0.39438229  3.82646648 
#>         349         350         351         352         353         354 
#>  1.28253499 -1.04890733  3.89831170  0.33626819  2.74797142  3.19244138 
#>         355         356         357         358         359         360 
#>  1.64762291  6.51241753 -1.61006877  0.22423441 -0.50784722  1.16465912 
#>         361         362         363         364         365         366 
#>  1.27894778  1.04518863 -1.01019178  0.25900287  3.56986386  5.20717962 
#>         367         368         369         370         371         372 
#>  0.07792432 -2.98638246 -2.30606705 -2.30051220 -1.21611471 -1.73288547 
#>         373         374         375         376         377         378 
#> -0.12169929 -0.83515159  0.01245007 -3.45273066 -1.65592973  1.05787762 
#>         379         380         381         382         383         384 
#>  0.93571959  2.59437694  2.01299577 -0.70142914  2.71225829  0.17290315 
#>         385         386         387         388         389         390 
#> -2.10906333  1.78347699 -1.25065173  5.11330745  0.79261923 -2.90228077 
#>         391         392         393         394         395         396 
#>  1.59745968  0.54060047  0.39000563  4.69943066  1.08203558 -1.62607548 
#>         397         398         399         400         401         402 
#>  1.18525682 -0.01087346  0.50793279  0.64680303  2.31838655  0.68934180 
#>         403         404         405         406         407         408 
#>  2.97664247  2.83790993  2.32888952  3.74530172 -1.84679799  6.19142774 
#>         409         410         411         412         413         414 
#> -1.11921219 -0.52027913 -1.84901917  3.90984792 -1.59586698  1.06946587 
#>         415         416         417         418         419         420 
#>  0.89897768 -0.75504304  1.38380983 -1.13768782  0.92300611  0.85960056 
#>         421         422         423         424         425         426 
#> -1.60919290  1.97287623  0.66399161 -1.47077555 -2.76413454  0.20813344 
#>         427         428         429         430         431         432 
#>  2.42928027  2.49096815 -0.64772450  1.03831859  0.11302629  0.79970802 
#>         433         434         435         436         437         438 
#>  0.70388983  1.81184152  3.48951877 -0.40923465  2.11208039  1.36334983 
#>         439         440         441         442         443         444 
#>  1.25178029 -0.72214603  5.85415534  0.03865567 -3.56142458  0.37092371 
#>         445         446         447         448         449         450 
#> -1.25763565 -1.19958852 -1.04920661  2.97513076 -3.40994906  3.16143738 
#>         451         452         453         454         455         456 
#> -0.70323365 -2.94644036 -1.42424053  0.61607910 -2.92373888 -1.37040177 
#>         457         458         459         460         461         462 
#>  3.52777801  0.37180818 -0.92658563  4.03531564 -1.99506616  2.09082851 
#>         463         464         465         466         467         468 
#> -2.08834752 -1.44188813 -2.65546149 -0.64120408  2.19455164 -1.30672605 
#>         469         470         471         472         473         474 
#> -0.55076083 -1.18548661  1.01117803  1.00685999  4.04427052  1.24667049 
#>         475         476         477         478         479         480 
#> -1.13931060  2.72901638  0.50919849  2.86579554  4.10615871  1.86239076 
#>         481         482         483         484         485         486 
#>  0.85566874 -2.53397985  1.93384041  0.34173635 -0.19896890 -3.17770038 
#>         487         488         489         490         491         492 
#> -0.10476977 -1.74397132 -2.83259881  5.46840525  1.86209966  2.45092697 
#>         493         494         495         496         497         498 
#>  1.00382366 -2.31638704 -2.73053947  0.26875918  0.33287075  0.41510084 
#>         499         500 
#> -0.02971605 -0.01738831 
predict(logistic_beta_x, model_type = "raw")
#>           1           2           3           4           5           6 
#> -0.14978605 -0.08134324 -0.17234895 -0.96500507 -0.41140099 -3.66500876 
#>           7           8           9          10          11          12 
#>  0.45641136  3.21533738  1.27031906  4.19968995 -1.12443124 -2.48912251 
#>          13          14          15          16          17          18 
#> -0.95030270  1.40075560 -3.87061892 -1.22196386  1.64793248  0.26351062 
#>          19          20          21          22          23          24 
#>  0.72796937 -3.10347290  2.51888487  2.95925773 -0.07194538  4.00420785 
#>          25          26          27          28          29          30 
#>  4.32166278  3.61108912  0.19253103  0.70736943  4.93483135  1.63179563 
#>          31          32          33          34          35          36 
#> -1.53121733  0.88326050 -1.56874072 -1.29975147 -2.31172710  0.80970723 
#>          37          38          39          40          41          42 
#> -2.74222060  2.76830583  0.14277124 -1.07515082  1.21997917 -3.56021526 
#>          43          44          45          46          47          48 
#> -0.37411008 -0.73927584 -0.32588437  2.92756024  3.48586073  0.79603101 
#>          49          50          51          52          53          54 
#>  0.13479416  2.15517198 -1.30365486 -0.97602940 -0.58075477 -0.10698828 
#>          55          56          57          58          59          60 
#> -2.79198254  2.97961397  0.34591136  0.64047900  2.92044253 -0.61555969 
#>          61          62          63          64          65          66 
#> -2.78373829 -3.17248201  1.04435794 -2.13417300 -3.12605466  1.46695915 
#>          67          68          69          70          71          72 
#> -0.27629999  0.06464732 -1.74345959  1.55472572  0.08787974 -3.95342379 
#>          73          74          75          76          77          78 
#> -4.48141414 -1.46129296  0.54234905  0.82867871 -0.32387428 -2.76947603 
#>          79          80          81          82          83          84 
#> -3.28820957 -1.10684008 -1.59519351 -2.15464871  1.14737887  1.12557393 
#>          85          86          87          88          89          90 
#>  2.75849747 -3.73809474 -1.33946564 -0.43821183  0.08230553 -2.66283963 
#>          91          92          93          94          95          96 
#> -2.22750598  2.04334594 -0.26442030 -0.17588281 -2.40653927  2.12545793 
#>          97          98          99         100         101         102 
#> -1.65670574  0.34246420  1.84471157 -3.41535617 -2.01583249  3.27498398 
#>         103         104         105         106         107         108 
#>  0.74220379  0.70309052 -1.71766823  0.08033343 -2.07154702 -1.00077386 
#>         109         110         111         112         113         114 
#> -2.23924613  2.67443454 -1.06628380 -0.25772980  1.66808167  1.44032300 
#>         115         116         117         118         119         120 
#> -1.19493355  0.60859601  0.14593548  1.04389437 -0.51851066  1.52205280 
#>         121         122         123         124         125         126 
#>  0.45109349  0.12090314 -3.46088450 -1.33896658  2.09200021  2.66663307 
#>         127         128         129         130         131         132 
#> -0.61582272 -2.45552247  1.05135084 -0.16159274  1.43234487  3.03601676 
#>         133         134         135         136         137         138 
#> -0.37651683  0.10070517  1.96454151  0.81307308  2.15618538  4.31247034 
#>         139         140         141         142         143         144 
#> -2.15053058 -2.52292605  2.75028684  1.73362226 -2.09302975 -2.60597462 
#>         145         146         147         148         149         150 
#>  0.72538904 -3.11278912  1.72165906 -1.64042493 -0.77554629  2.39461855 
#>         151         152         153         154         155         156 
#> -1.70863446 -0.17428506 -0.82468356  0.06074132 -1.16490507  0.29509604 
#>         157         158         159         160         161         162 
#> -0.68398793 -2.05277591  2.61431096  2.10334114 -0.52508101  3.01338466 
#>         163         164         165         166         167         168 
#>  0.11406477  2.77784767  1.54997586 -1.10641586 -0.44918478  2.91835191 
#>         169         170         171         172         173         174 
#>  1.47430119  0.99684868  0.49947626  3.12005526  0.48128938  2.77794721 
#>         175         176         177         178         179         180 
#>  0.16595507  0.16861120 -0.22376824 -1.31872945  1.85209254 -2.54602319 
#>         181         182         183         184         185         186 
#> -2.53859081 -1.88972430 -3.76687457  3.70116151 -3.21395851  2.89085930 
#>         187         188         189         190         191         192 
#> -1.40479191 -4.29621824  0.67899091  1.89972660  0.24971668 -1.79169593 
#>         193         194         195         196         197         198 
#> -2.12379832 -0.58494100 -0.27143124  0.32580796  4.84565233 -1.31177570 
#>         199         200         201         202         203         204 
#> -3.06889586  0.60330544 -0.92431558 -1.37087778  0.18234070 -0.23410712 
#>         205         206         207         208         209         210 
#> -1.72683579  2.97224220 -3.83487255 -1.59774187  3.09591749 -2.83625462 
#>         211         212         213         214         215         216 
#>  0.57501931 -1.18626605 -1.29232506  2.66235630  2.54333553  2.23068544 
#>         217         218         219         220         221         222 
#> -1.86052114  2.73540792  1.60009735  0.40208225 -2.76027561 -0.66387341 
#>         223         224         225         226         227         228 
#> -0.83102897 -1.67344023  4.13435098 -1.31790519 -0.40528978 -2.10472583 
#>         229         230         231         232         233         234 
#> -1.35923972  2.74534028  1.67655717  2.31455720  4.87817883  0.99003311 
#>         235         236         237         238         239         240 
#> -0.93981548  1.75345796 -1.75439883 -2.22564108  0.07093593  0.37916963 
#>         241         242         243         244         245         246 
#> -1.48224979  3.21970629 -1.23280102 -3.86674391  0.61502781  1.50965136 
#>         247         248         249         250         251         252 
#> -0.30348683 -2.50054556  1.37955219 -2.26703558 -2.55422215  1.72805377 
#>         253         254         255         256         257         258 
#> -0.49886308 -0.40496537  1.60535580 -2.34509376 -1.56954013 -2.76911255 
#>         259         260         261         262         263         264 
#> -1.84182565  2.11649198 -6.15055823 -0.55362932 -2.27302723  5.72724308 
#>         265         266         267         268         269         270 
#> -0.76130330 -3.99726268  0.32848350  1.01031728  0.83868488 -2.13292685 
#>         271         272         273         274         275         276 
#>  2.21734020  1.99475004 -1.56410453  1.31986166 -0.24876337  3.83256471 
#>         277         278         279         280         281         282 
#>  0.77100504  2.71582991  2.42387504  1.64199526  1.70696492 -3.21105147 
#>         283         284         285         286         287         288 
#> -2.34701803 -0.04934355  3.61706269  0.52696545  0.09540556 -4.28534988 
#>         289         290         291         292         293         294 
#> -1.82039761 -2.51962413 -0.24888060 -0.11925439 -2.03556616  2.81106731 
#>         295         296         297         298         299         300 
#> -1.50652775 -1.65268211 -0.22805644 -1.02193963 -0.06059586  1.20082000 
#>         301         302         303         304         305         306 
#>  4.01727763  0.56872269  4.78872141 -1.31901837 -0.63426880  1.89050136 
#>         307         308         309         310         311         312 
#> -1.72599666  4.27074031  0.51840777 -1.72449479 -3.26146144  1.08686025 
#>         313         314         315         316         317         318 
#>  0.81644387  1.71325127 -1.93810249 -2.65080675 -0.79585912 -3.49791403 
#>         319         320         321         322         323         324 
#>  2.40564389 -0.41507462  5.37568060 -1.45458564 -1.06606591 -0.63344740 
#>         325         326         327         328         329         330 
#>  2.88906242  0.72274356 -2.87207798 -1.11506590  0.73601529 -1.64994148 
#>         331         332         333         334         335         336 
#>  2.62882867 -2.05228099  1.62367184 -0.40381925  2.19622707 -0.41270011 
#>         337         338         339         340         341         342 
#>  1.34802157  0.74418499  1.98198510  1.90776354 -0.97432347 -0.18827959 
#>         343         344         345         346         347         348 
#> -1.23045646  1.78320427 -1.10338674 -3.36752264  0.39438229  3.82646648 
#>         349         350         351         352         353         354 
#>  1.28253499 -1.04890733  3.89831170  0.33626819  2.74797142  3.19244138 
#>         355         356         357         358         359         360 
#>  1.64762291  6.51241753 -1.61006877  0.22423441 -0.50784722  1.16465912 
#>         361         362         363         364         365         366 
#>  1.27894778  1.04518863 -1.01019178  0.25900287  3.56986386  5.20717962 
#>         367         368         369         370         371         372 
#>  0.07792432 -2.98638246 -2.30606705 -2.30051220 -1.21611471 -1.73288547 
#>         373         374         375         376         377         378 
#> -0.12169929 -0.83515159  0.01245007 -3.45273066 -1.65592973  1.05787762 
#>         379         380         381         382         383         384 
#>  0.93571959  2.59437694  2.01299577 -0.70142914  2.71225829  0.17290315 
#>         385         386         387         388         389         390 
#> -2.10906333  1.78347699 -1.25065173  5.11330745  0.79261923 -2.90228077 
#>         391         392         393         394         395         396 
#>  1.59745968  0.54060047  0.39000563  4.69943066  1.08203558 -1.62607548 
#>         397         398         399         400         401         402 
#>  1.18525682 -0.01087346  0.50793279  0.64680303  2.31838655  0.68934180 
#>         403         404         405         406         407         408 
#>  2.97664247  2.83790993  2.32888952  3.74530172 -1.84679799  6.19142774 
#>         409         410         411         412         413         414 
#> -1.11921219 -0.52027913 -1.84901917  3.90984792 -1.59586698  1.06946587 
#>         415         416         417         418         419         420 
#>  0.89897768 -0.75504304  1.38380983 -1.13768782  0.92300611  0.85960056 
#>         421         422         423         424         425         426 
#> -1.60919290  1.97287623  0.66399161 -1.47077555 -2.76413454  0.20813344 
#>         427         428         429         430         431         432 
#>  2.42928027  2.49096815 -0.64772450  1.03831859  0.11302629  0.79970802 
#>         433         434         435         436         437         438 
#>  0.70388983  1.81184152  3.48951877 -0.40923465  2.11208039  1.36334983 
#>         439         440         441         442         443         444 
#>  1.25178029 -0.72214603  5.85415534  0.03865567 -3.56142458  0.37092371 
#>         445         446         447         448         449         450 
#> -1.25763565 -1.19958852 -1.04920661  2.97513076 -3.40994906  3.16143738 
#>         451         452         453         454         455         456 
#> -0.70323365 -2.94644036 -1.42424053  0.61607910 -2.92373888 -1.37040177 
#>         457         458         459         460         461         462 
#>  3.52777801  0.37180818 -0.92658563  4.03531564 -1.99506616  2.09082851 
#>         463         464         465         466         467         468 
#> -2.08834752 -1.44188813 -2.65546149 -0.64120408  2.19455164 -1.30672605 
#>         469         470         471         472         473         474 
#> -0.55076083 -1.18548661  1.01117803  1.00685999  4.04427052  1.24667049 
#>         475         476         477         478         479         480 
#> -1.13931060  2.72901638  0.50919849  2.86579554  4.10615871  1.86239076 
#>         481         482         483         484         485         486 
#>  0.85566874 -2.53397985  1.93384041  0.34173635 -0.19896890 -3.17770038 
#>         487         488         489         490         491         492 
#> -0.10476977 -1.74397132 -2.83259881  5.46840525  1.86209966  2.45092697 
#>         493         494         495         496         497         498 
#>  1.00382366 -2.31638704 -2.73053947  0.26875918  0.33287075  0.41510084 
#>         499         500 
#> -0.02971605 -0.01738831