Standardize Coefficients in a 'lavaan'-Model

Can standardize selected variables in a lavaan model without refitting the models, can handle product term correctly and skip categorical predictors in standardization.

Usage

lav_betaselect(
  object,
  to_standardize = ".all.",
  not_to_standardize = NULL,
  skip_categorical_x = TRUE,
  output = c("data.frame", "text"),
  std_se = c("none", "delta", "bootstrap"),
  std_z = TRUE,
  std_pvalue = TRUE,
  std_ci = TRUE,
  level = 0.95,
  progress = TRUE,
  boot_out = NULL,
  bootstrap = 100L,
  store_boot_est = TRUE,
  parallel = c("no", "snow", "multicore"),
  ncpus = parallel::detectCores(logical = FALSE) - 1,
  cl = NULL,
  iseed = NULL,
  find_product_terms = TRUE,
  ...,
  delta_method = c("lavaan", "numDeriv"),
  vector_form = TRUE
)

Arguments

object

The output of lavaan model fit functions, such as lavaan::sem() and lavaan::cfa().

to_standardize

A string vector, which should be the names of the variables to be standardized. Default is ".all", indicating all variables are to be standardized (but see skip_categorical).

not_to_standardize

A string vector, which should be the names of the variables that should not be standardized. This argument is useful when most variables, except for a few, are to be standardized. This argument cannot be ued with to_standardize at the same time. Default is NULL, and only to_standardize is used.

skip_categorical_x

Logical. If TRUE, the default, all categorical predictors, defined as variables with only two possible values in the data analyzed, will be skipped in standardization. This overrides the argument to_standardize. That is, a categorical predictor will not be standardized even if listed in to_standardize, unless uses set this argument to FALSE.

output

The format of the output. If set to "data.frame", the output will be a data frame like the one from lavaan::parameterEstimates(). If set to "text", it will be printed in a format similar to the one in the summary-method of lavaan. (NOTE: "text" supported but not fully implemented.)

std_se

String. If set to "none", the default, standard errors will not be computed for the standardized solution. If set to "delta", delta method will be used to compute the standard errors. If set to "bootstrap", then what it does depends whether boot_out is set. If boot_out is to an output of manymome::do_boot(), its content will be used. If boot_out is NULL and bootstrap estimates are available in object (e.g., bootstrapping is requested when fitting the model in lavaan), then the stored bootstrap estimates will be sued. If not available, the bootstrapping will be conducted using lavaan::bootstrapLavaan(), using arguments bootstrap, parallel, ncpus, cl, and iseed.`

std_z

Logical. If TRUE and std_se is not set to "none", standard error will be computed using the method specified in std_se. Default is TRUE.

std_pvalue

Logical. If TRUE, std_se is not set to "none", and std_z is TRUE, p-values will be computed using the method specified in std_se. For bootstrapping, the method proposed by Asparouhov and Muthén (2021) is used. Default is TRUE.

std_ci

Logical. If TRUE and std_se is not set to "none", confidence intervals will be computed using the method specified in std_se. Default is FALSE.

level

The level of confidence of the confidence intervals. Default is .95. It will be used in the confidence intervals of both the unstandardized and standardized solution.

progress

Logical. If TRUE, progress bars will be displayed for long process.

boot_out

If std_se is "bootstrap" and this argument is set to an output of manymome::do_boot(), its output will be used in computing statistics such as standard errors and confidence intervals. This allows users to use methods other than bootstrapping when fitting the model, while they can still request bootstrapping for the standardized solution.

bootstrap

If std_se is "bootstrap" but bootstrapping is not requested when fitting the model and boot_out is not set, lavaan::bootstrapLavaan() will be called to do bootstrapping. This argument is the number of bootstrap samples to draw. Default is 100. Should be set to 5000 or even 10000 for stable results.

store_boot_est

Logical. If std_se is "bootstrap" and this argument is TRUE, the default, the bootstrap estimates of the standardized solution will be stored in the attribute "boot_est". These estimates can be used for diagnosis of the bootstrapping. If FALSE, then the bootstrap estimates will not be stored.

parallel

If std_se is "bootstrap" but bootstrapping is not requested when fitting the model and boot_out is not set, lavaan::bootstrapLavaan() will be called to do bootstrapping. This argument is to be passed to lavaan::bootstrapLavaan(). Default is "no".

ncpus

If std_se is "bootstrap" but bootstrapping is not requested when fitting the model and boot_out is not set, lavaan::bootstrapLavaan() will be called to do bootstrapping. This argument is to be passed to lavaan::bootstrapLavaan(). Default is parallel::detectCores(logical = FALSE) - 1. Ignored if parallel is "no".

cl

If std_se is "bootstrap" but bootstrapping is not requested when fitting the model and boot_out is not set, lavaan::bootstrapLavaan() will be called to do bootstrapping. This argument is to be passed to lavaan::bootstrapLavaan(). Default is NULL. Ignored if parallel is "no".

iseed

If std_se is "bootstrap" but bootstrapping is not requested when fitting the model and boot_out is not set, lavaan::bootstrapLavaan() will be called to do bootstrapping. This argument is to be passed to lavaan::bootstrapLavaan() to set the seed for the random resampling. Default is NULL. Should be set to an integer for reproducible results. Ignored if parallel is "no".

find_product_terms

String. If it is certain that a model does not have product terms, setting this to FALSE will skip the search, which is time consuming for a models with many paths and/or many variables. Default is TRUE, and the function will automatically identify product terms, if any.

...

Optional arguments to be passed to the lavaan::parameterEstimates(), which will be use to generate the output.

delta_method

The method used to compute delta method standard errors. For internal use and should not be changed.

vector_form

The internal method used to compute standardized solution. For internal use and should not be changed.

Value

A data frame storing the parameter estimates, similar in form to the output of lavaan::parameterEstimates().

Details

This function lets users select which variables to be standardized when computing the standardized solution. It has the following features:

It automatically skips predictors which has only two unique values, assuming that they are dummy variables.

It does not standardize product term, which is incorrect. Instead, it compute the product term with its component variables standardized.

It can be used to generate bootstrap confidence intervals for the standardized solution. Bootstrap confidence interval is better than doing standardization before fitting a model because it correctly takes into account the sampling variance of the standard deviations. It is also better than delta method confidence interval because it takes into account the usually asymmetric distribution of parameters after standardization.

For comparison, it can also report delta method standard errors and confidence intervals.

Problems With Common Approaches

In most SEM programs, users have limited control on which variables to standardize when requesting the standardized solution. The solution may be uninterpretable or misleading in these conditions:

Dummy variables are standardized and can not be interpreted as the difference between two groups on the outcome variables.

Product terms (interaction terms) are standardized and they cannot be interpreted as the changes in the effects of focal variables when the moderators change (Cheung, Cheung, Lau, Hui, & Vong, 2022).

Variables with meaningful units can be more difficult to interpret when they are standardized (e.g., age).

Moreover, the delta method is usually used, which is suboptimal for standardization unless the sample size is large. For example, the covariance with variables standardized is a correlation, and its sampling distribution is skewed unless its population value is zero. However, delta method confidence interval for the correlation is necessarily symmetric around the point estimate.

Limitations

It only supports observed variable interaction terms, and only support two-way interactions.

It does not support multilevel models.

It only supports models fitted to raw data.

Intercepts not supported.

References

Asparouhov, A., & Muthén, B. (2021). Bootstrap p-value computation. Retrieved from https://www.statmodel.com/download/FAQ-Bootstrap%20-%20Pvalue.pdf

Cheung, S. F., Cheung, S.-H., Lau, E. Y. Y., Hui, C. H., & Vong, W. N. (2022) Improving an old way to measure moderation effect in standardized units. Health Psychology, 41(7), 502-505. doi:10.1037/hea0001188

See also

Author

Shu Fai Cheung https://orcid.org/0000-0002-9871-9448

Examples

library(lavaan)
mod <-
"
med ~ iv + mod + iv:mod
dv ~ med + iv
"
fit <- sem(mod,
           data_test_medmod,
           fixed.x = TRUE)
summary(fit)
#> lavaan 0.6-18 ended normally after 3 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of model parameters                         7
#> 
#>   Number of observations                           200
#> 
#> Model Test User Model:
#>                                                       
#>   Test statistic                                 2.685
#>   Degrees of freedom                                 2
#>   P-value (Chi-square)                           0.261
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Standard
#>   Information                                 Expected
#>   Information saturated (h1) model          Structured
#> 
#> Regressions:
#>                    Estimate   Std.Err  z-value  P(>|z|)
#>   med ~                                                
#>     iv                -6.339    0.997   -6.357    0.000
#>     mod               -3.903    0.622   -6.277    0.000
#>     iv:mod             0.286    0.039    7.248    0.000
#>   dv ~                                                 
#>     med                0.093    0.011    8.298    0.000
#>     iv                 0.229    0.039    5.917    0.000
#> 
#> Variances:
#>                    Estimate   Std.Err  z-value  P(>|z|)
#>    .med               61.851    6.185   10.000    0.000
#>    .dv                 2.104    0.210   10.000    0.000
#> 
fit_beta <- lav_betaselect(fit,
                           to_standardize = c("iv", "dv"))
fit_beta
#>       lhs op    rhs       est    se      z pvalue  ci.lower  ci.upper    std.p
#> 1     med  ~     iv    -6.339 0.997 -6.357      0    -8.293    -4.385  -17.697
#> 2     med  ~    mod    -3.903 0.622 -6.277      0    -5.122    -2.684   -3.903
#> 3     med  ~ iv:mod     0.286 0.039  7.248      0     0.208     0.363    0.797
#> 4      dv  ~    med     0.093 0.011  8.298      0     0.071     0.115    0.049
#> 5      dv  ~     iv     0.229 0.039  5.917      0     0.153     0.304    0.333
#> 6     med ~~    med    61.851 6.185 10.000      0    49.728    73.974   61.851
#> 7      dv ~~     dv     2.104 0.210 10.000      0     1.692     2.517    0.574
#> 8      iv ~~     iv     7.795 0.000     NA     NA     7.795     7.795    1.000
#> 9      iv ~~    mod     5.287 0.000     NA     NA     5.287     5.287    1.894
#> 10     iv ~~ iv:mod   274.291 0.000     NA     NA   274.291   274.291   35.189
#> 11    mod ~~    mod    23.129 0.000     NA     NA    23.129    23.129   23.129
#> 12    mod ~~ iv:mod   488.157 0.000     NA     NA   488.157   488.157  174.847
#> 13 iv:mod ~~ iv:mod 14521.504 0.000     NA     NA 14521.504 14521.504 1862.983
#>    std.p.by
#> 1        iv
#> 2          
#> 3        iv
#> 4        dv
#> 5     iv,dv
#> 6          
#> 7        dv
#> 8        iv
#> 9        iv
#> 10       iv
#> 11         
#> 12       iv
#> 13       iv
print(fit_beta,
      output = "text")
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Standard
#>   Information                                 Expected
#>   Information saturated (h1) model          Structured
#> 
#> Selected Standardization (Bs):
#> 
#>   Only selected variables are standardized
#> 
#> Regressions:
#>                    Estimate   Std.Err  z-value  P(>|z|)    CI.Lo     CI.Hi 
#>   med ~                                                                    
#>     iv                -6.339    0.997   -6.357    0.000    -8.293    -4.385
#>     mod               -3.903    0.622   -6.277    0.000    -5.122    -2.684
#>     iv:mod             0.286    0.039    7.248    0.000     0.208     0.363
#>   dv ~                                                                     
#>     med                0.093    0.011    8.298    0.000     0.071     0.115
#>     iv                 0.229    0.039    5.917    0.000     0.153     0.304
#>        Bs
#>          
#>   -17.697
#>    -3.903
#>     0.797
#>          
#>     0.049
#>     0.333
#> 
#> Variances:
#>                    Estimate   Std.Err  z-value  P(>|z|)    CI.Lo     CI.Hi 
#>    .med               61.851    6.185   10.000    0.000    49.728    73.974
#>    .dv                 2.104    0.210   10.000    0.000     1.692     2.517
#>        Bs
#>    61.851
#>     0.574
#> 
#> Note:
#> - These variables are standardized: iv, dv
#> - Product terms (iv:mod) have variables standardized before computing
#>   them. That is, the product terms themselves are not standardized.

# In real studies:
# - should set bootstrap to at least 5000
# - should set parallel to "snow" or "multicore"
fit_beta_boot <- lav_betaselect(fit,
                                to_standardize = c("iv", "dv"),
                                std_se = "bootstrap",
                                std_ci = TRUE,
                                bootstrap = 100,
                                iseed = 1234)
fit_beta_boot
#>       lhs op    rhs       est    se      z pvalue  ci.lower  ci.upper    std.p
#> 1     med  ~     iv    -6.339 0.997 -6.357      0    -8.293    -4.385  -17.697
#> 2     med  ~    mod    -3.903 0.622 -6.277      0    -5.122    -2.684   -3.903
#> 3     med  ~ iv:mod     0.286 0.039  7.248      0     0.208     0.363    0.797
#> 4      dv  ~    med     0.093 0.011  8.298      0     0.071     0.115    0.049
#> 5      dv  ~     iv     0.229 0.039  5.917      0     0.153     0.304    0.333
#> 6     med ~~    med    61.851 6.185 10.000      0    49.728    73.974   61.851
#> 7      dv ~~     dv     2.104 0.210 10.000      0     1.692     2.517    0.574
#> 8      iv ~~     iv     7.795 0.000     NA     NA     7.795     7.795    1.000
#> 9      iv ~~    mod     5.287 0.000     NA     NA     5.287     5.287    1.894
#> 10     iv ~~ iv:mod   274.291 0.000     NA     NA   274.291   274.291   35.189
#> 11    mod ~~    mod    23.129 0.000     NA     NA    23.129    23.129   23.129
#> 12    mod ~~ iv:mod   488.157 0.000     NA     NA   488.157   488.157  174.847
#> 13 iv:mod ~~ iv:mod 14521.504 0.000     NA     NA 14521.504 14521.504 1862.983
#>    std.p.by std.p.se std.p.z std.p.pvalue std.p.ci.lower std.p.ci.upper
#> 1        iv    2.856  -6.196            0        -22.798        -11.374
#> 2              0.682  -5.724            0         -5.102         -2.513
#> 3        iv    0.114   6.987            0          0.567          1.018
#> 4        dv    0.005  10.592            0          0.038          0.057
#> 5     iv,dv    0.046   7.179            0          0.226          0.412
#> 6              5.836  10.598            0         48.732         72.713
#> 7        dv    0.054  10.695            0          0.441          0.684
#> 8        iv    0.000      NA            0          1.000          1.000
#> 9        iv    0.000      NA            0          1.894          1.894
#> 10       iv    0.000      NA            0         35.189         35.189
#> 11             0.000      NA            0         23.129         23.129
#> 12       iv    0.000      NA            0        174.847        174.847
#> 13       iv    0.000      NA            0       1862.983       1862.983
print(fit_beta_boot,
      output = "text")
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Standard
#>   Information                                 Expected
#>   Information saturated (h1) model          Structured
#> 
#> Selected Standardization (Bs):
#> 
#>   Only selected variables are standardized
#>   Standard errors                            Bootstrap
#>   Bootstrap samples                                100
#> 
#> Regressions:
#>                    Estimate   Std.Err  z-value  P(>|z|)    CI.Lo     CI.Hi 
#>   med ~                                                                    
#>     iv                -6.339    0.997   -6.357    0.000    -8.293    -4.385
#>     mod               -3.903    0.622   -6.277    0.000    -5.122    -2.684
#>     iv:mod             0.286    0.039    7.248    0.000     0.208     0.363
#>   dv ~                                                                     
#>     med                0.093    0.011    8.298    0.000     0.071     0.115
#>     iv                 0.229    0.039    5.917    0.000     0.153     0.304
#>        Bs    Bs.se     Bs.z Bs.pvalue Bs.CI.Lo Bs.CI.Hi
#>                                                        
#>   -17.697    2.856   -6.196    0.000   -22.798  -11.374
#>    -3.903    0.682   -5.724    0.000    -5.102   -2.513
#>     0.797    0.114    6.987    0.000     0.567    1.018
#>                                                        
#>     0.049    0.005   10.592    0.000     0.038    0.057
#>     0.333    0.046    7.179    0.000     0.226    0.412
#> 
#> Variances:
#>                    Estimate   Std.Err  z-value  P(>|z|)    CI.Lo     CI.Hi 
#>    .med               61.851    6.185   10.000    0.000    49.728    73.974
#>    .dv                 2.104    0.210   10.000    0.000     1.692     2.517
#>        Bs    Bs.se     Bs.z Bs.pvalue Bs.CI.Lo Bs.CI.Hi
#>    61.851    5.836   10.598    0.000    48.732   72.713
#>     0.574    0.054   10.695    0.000     0.441    0.684
#> 
#> Note:
#> - These variables are standardized: iv, dv
#> - Product terms (iv:mod) have variables standardized before computing
#>   them. That is, the product terms themselves are not standardized.

# Print only beta_selects
print(fit_beta_boot,
      output = "text",
      standardized_only = TRUE)
#> 
#> Selected Standardized Estimates Only (Bs):
#> 
#>   Only selected variables are standardized
#>   Standard errors                            Bootstrap
#>   Bootstrap samples                                100
#> 
#> Regressions:
#>           Standardized (Bs)  Std.Err  z-value  p-value ci.lower ci.upper
#>   med ~                                                                 
#>     iv              -17.697    2.856   -6.196    0.000  -22.798  -11.374
#>     mod              -3.903    0.682   -5.724    0.000   -5.102   -2.513
#>     iv:mod            0.797    0.114    6.987    0.000    0.567    1.018
#>   dv ~                                                                  
#>     med               0.049    0.005   10.592    0.000    0.038    0.057
#>     iv                0.333    0.046    7.179    0.000    0.226    0.412
#> 
#> Variances:
#>           Standardized (Bs)  Std.Err  z-value  p-value ci.lower ci.upper
#>    .med              61.851    5.836   10.598    0.000   48.732   72.713
#>    .dv                0.574    0.054   10.695    0.000    0.441    0.684
#> 
#> Note:
#> - These variables are standardized: iv, dv
#> - Product terms (iv:mod) have variables standardized before computing
#>   them. That is, the product terms themselves are not standardized.