Print a 'lav_betaselect' Object

Print method for an 'lav_betaselect' object, which is the output of lav_betaselect().

Usage

# S3 method for lav_betaselect
print(x, ..., nd = 3, output = c("table", "text"), standardized_only = FALSE)

Arguments

x

Object of the class std_solution_boot.

...

Optional arguments to be passed to print() methods.

nd

The number of digits after the decimal place. Default is 3.

output

String. How the results are printed. Default is "table" and the results are printed in a table format similar to that of lavaan::parameterEstimates() with output set to "table". If "text", the results will be printed in a text format similar to the printout of the output of summary() of a lavaan object.

standardized_only

Logical. If TRUE, only the results for the standardized solution will be printed. If FALSE, the default, then the standardized solution is printed alongside the unstandardized solution, as in the printout of the output of summary() of a lavaan object.

Value

x is returned invisibly. Called for its side effect.

Details

The default format of the printout is that of lavaan::parameterEstimates(), which is compact but not easy to read. Users can request a format similar to that of the printout of the summary of a lavaan output by setting output to "text".

For the "text" format, users can also select whether only the standardized solution is printed or whether the standardized solution is appended to the right of the printout, as in the output of summary() for a lavaan output.

See also

Author

Shu Fai Cheung https://orcid.org/0000-0002-9871-9448

Examples

library(lavaan)
mod <-
"
med ~ iv + mod + iv:mod
dv ~ med + iv
"
fit <- sem(mod,
           data_test_medmod,
           fixed.x = TRUE)
summary(fit)
#> lavaan 0.6-18 ended normally after 3 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of model parameters                         7
#> 
#>   Number of observations                           200
#> 
#> Model Test User Model:
#>                                                       
#>   Test statistic                                 2.685
#>   Degrees of freedom                                 2
#>   P-value (Chi-square)                           0.261
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Standard
#>   Information                                 Expected
#>   Information saturated (h1) model          Structured
#> 
#> Regressions:
#>                    Estimate   Std.Err  z-value  P(>|z|)
#>   med ~                                                
#>     iv                -6.339    0.997   -6.357    0.000
#>     mod               -3.903    0.622   -6.277    0.000
#>     iv:mod             0.286    0.039    7.248    0.000
#>   dv ~                                                 
#>     med                0.093    0.011    8.298    0.000
#>     iv                 0.229    0.039    5.917    0.000
#> 
#> Variances:
#>                    Estimate   Std.Err  z-value  P(>|z|)
#>    .med               61.851    6.185   10.000    0.000
#>    .dv                 2.104    0.210   10.000    0.000
#> 
fit_beta <- lav_betaselect(fit,
                           to_standardize = c("iv", "dv"))
fit_beta
#>       lhs op    rhs       est    se      z pvalue  ci.lower  ci.upper    std.p
#> 1     med  ~     iv    -6.339 0.997 -6.357      0    -8.293    -4.385  -17.697
#> 2     med  ~    mod    -3.903 0.622 -6.277      0    -5.122    -2.684   -3.903
#> 3     med  ~ iv:mod     0.286 0.039  7.248      0     0.208     0.363    0.797
#> 4      dv  ~    med     0.093 0.011  8.298      0     0.071     0.115    0.049
#> 5      dv  ~     iv     0.229 0.039  5.917      0     0.153     0.304    0.333
#> 6     med ~~    med    61.851 6.185 10.000      0    49.728    73.974   61.851
#> 7      dv ~~     dv     2.104 0.210 10.000      0     1.692     2.517    0.574
#> 8      iv ~~     iv     7.795 0.000     NA     NA     7.795     7.795    1.000
#> 9      iv ~~    mod     5.287 0.000     NA     NA     5.287     5.287    1.894
#> 10     iv ~~ iv:mod   274.291 0.000     NA     NA   274.291   274.291   35.189
#> 11    mod ~~    mod    23.129 0.000     NA     NA    23.129    23.129   23.129
#> 12    mod ~~ iv:mod   488.157 0.000     NA     NA   488.157   488.157  174.847
#> 13 iv:mod ~~ iv:mod 14521.504 0.000     NA     NA 14521.504 14521.504 1862.983
#>    std.p.by
#> 1        iv
#> 2          
#> 3        iv
#> 4        dv
#> 5     iv,dv
#> 6          
#> 7        dv
#> 8        iv
#> 9        iv
#> 10       iv
#> 11         
#> 12       iv
#> 13       iv
print(fit_beta,
      output = "text")
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Standard
#>   Information                                 Expected
#>   Information saturated (h1) model          Structured
#> 
#> Selected Standardization (Bs):
#> 
#>   Only selected variables are standardized
#> 
#> Regressions:
#>                    Estimate   Std.Err  z-value  P(>|z|)    CI.Lo     CI.Hi 
#>   med ~                                                                    
#>     iv                -6.339    0.997   -6.357    0.000    -8.293    -4.385
#>     mod               -3.903    0.622   -6.277    0.000    -5.122    -2.684
#>     iv:mod             0.286    0.039    7.248    0.000     0.208     0.363
#>   dv ~                                                                     
#>     med                0.093    0.011    8.298    0.000     0.071     0.115
#>     iv                 0.229    0.039    5.917    0.000     0.153     0.304
#>        Bs    Bs.by
#>                   
#>   -17.697    iv   
#>    -3.903         
#>     0.797    iv   
#>                   
#>     0.049    dv   
#>     0.333 iv,dv   
#> 
#> Variances:
#>                    Estimate   Std.Err  z-value  P(>|z|)    CI.Lo     CI.Hi 
#>    .med               61.851    6.185   10.000    0.000    49.728    73.974
#>    .dv                 2.104    0.210   10.000    0.000     1.692     2.517
#>        Bs    Bs.by
#>    61.851         
#>     0.574    dv   
#> 
#> Note:
#> - The column 'Bs.by' lists variable(s) standardized when computing the
#>   standardized coefficient of a parameter. ('NA' for user-defined
#>   parameters because they are computed from other standardized
#>   parameters.)
#> - These variables are standardized: iv, dv
#> - Product terms (iv:mod) have variables standardized before computing
#>   them. That is, the product terms themselves are not standardized.