The 'vcov' Method for an 'lm_betaselect'-Class Object

Compute the variance-covariance matrix of estimates in the output of lm_betaselect().

Usage

# S3 method for lm_betaselect
vcov(
  object,
  method = c("boot", "bootstrap", "ls"),
  type = c("beta", "standardized", "raw", "unstandardized"),
  warn = TRUE,
  ...
)

Arguments

object

The output of lm_betaselect() or an lm_betaselect-class object.

method

The method used to compute the variance-covariance matrix. If bootstrapping was requested when calling lm_betaselect() and this argument is set to "bootstrap" or "boot", the bootstrap variance-covariance matrix is returned. If bootstrapping was not requested or if this argument is set to "ls", then the usual lm variance-covariance matrix is returned, with a warning raised unless type is "raw" or "unstandardized". Default is "boot".

type

String. If "unstandardized" or "raw", the variance-covariance matrix of the coefficients before standardization are returned. If "beta" or "standardized", then the variance-covariance matrix of the coefficients after selected variables standardized are returned. Default is "beta".

warn

Logical. WHether a warning will be raised is OLS (or WLS) variance-covariance matrix is requested for the model with some variables standardized (i.e., type is "beta" or "standardized"). Default is TRUE.

...

Other arguments to be passed to stats::vcov().

Value

A matrix of the variances and covariances of the parameter estimates.

Details

The type of variance-covariance matrix depends on the object. If bootstrapping was requested, by default it returns the bootstrap variance-covariance matrix. Otherwise, it returns the OLS (or WLS) variance-covariance matrix and raises a warning.

Support for other type of variance-covariance matrix will be added.

Author

Shu Fai Cheung https://orcid.org/0000-0002-9871-9448

Examples

data(data_test_mod_cat)

# bootstrap should be set to 2000 or 5000 in real studies
lm_beta_x <- lm_betaselect(dv ~ iv*mod + cov1 + cat1,
                           data = data_test_mod_cat,
                           to_standardize = "iv",
                           do_boot = TRUE,
                           bootstrap = 100,
                           iseed = 1234)
vcov(lm_beta_x)
#>             (Intercept)          iv          mod        cov1      cat1gp2
#> (Intercept) 341529.1966 -30895.9870 -3464.959234  719.311869   6119.71357
#> iv          -30895.9870 390982.0777   418.709645 -934.754965 -11173.61930
#> mod          -3464.9592    418.7096    36.371280  -15.826380    -98.89705
#> cov1           719.3119   -934.7550   -15.826380   89.431977    110.16302
#> cat1gp2       6119.7136 -11173.6193   -98.897047  110.163020   5596.41208
#> cat1gp3        168.9982  -6374.1237   -27.706276  -11.015161   2379.59442
#> iv:mod         317.3652  -3869.0948    -4.181951    8.995417    104.91294
#>                 cat1gp3       iv:mod
#> (Intercept)   168.99821   317.365230
#> iv          -6374.12372 -3869.094815
#> mod           -27.70628    -4.181951
#> cov1          -11.01516     8.995417
#> cat1gp2      2379.59442   104.912942
#> cat1gp3      5405.70924    61.328872
#> iv:mod         61.32887    38.479356
vcov(lm_beta_x, method = "ls")
#> Warning: With standardization, the variance-covariance matrix using OLS or WLS should not be used.
#>               (Intercept)           iv           mod          cov1     cat1gp2
#> (Intercept) 384454.213030    855.64316 -3723.4966214 -1097.5954358  -632.86557
#> iv             855.643164 369545.84716    36.6391492  -270.3857355 -2377.70021
#> mod          -3723.496621     36.63915    37.3035605     0.8903597   -20.15980
#> cov1         -1097.595436   -270.38574     0.8903597   103.4952425   -33.43135
#> cat1gp2       -632.865575  -2377.70021   -20.1598049   -33.4313529  5783.03424
#> cat1gp3        474.235748  -1276.79274   -30.9821620   -35.2964697  2994.51496
#> iv:mod          -1.815315  -3681.99659    -0.4181725     2.6484043    22.01237
#>                 cat1gp3        iv:mod
#> (Intercept)   474.23575    -1.8153150
#> iv          -1276.79274 -3681.9965904
#> mod           -30.98216    -0.4181725
#> cov1          -35.29647     2.6484043
#> cat1gp2      2994.51496    22.0123669
#> cat1gp3      5643.94687    11.5915722
#> iv:mod         11.59157    36.7790296
vcov(lm_beta_x, type = "raw")
#>              (Intercept)            iv           mod        cov1     cat1gp2
#> (Intercept) 22377927.302 -1455331.0752 -224020.61098 7757.749242 89893.97909
#> iv          -1455331.075    96108.5731   14544.46771 -468.507751 -5592.14211
#> mod          -224020.611    14544.4677    2249.82334  -85.068231  -901.81123
#> cov1            7757.749     -468.5078     -85.06823   89.431977   110.16302
#> cat1gp2        89893.979    -5592.1421    -901.81123  110.163020  5596.41208
#> cat1gp3        49297.662    -3270.8129    -513.93637  -11.015161  2379.59442
#> iv:mod         14484.706     -955.2700    -145.18126    4.575244    53.41171
#>                 cat1gp3       iv:mod
#> (Intercept) 49297.66173 14484.705612
#> iv          -3270.81287  -955.269975
#> mod          -513.93637  -145.181261
#> cov1          -11.01516     4.575244
#> cat1gp2      2379.59442    53.411713
#> cat1gp3      5405.70924    32.167824
#> iv:mod         32.16782     9.524374